+++ src/discljord/messaging.clj
@@ -577,3 +577,8 @@
   ""
   [webhook-token]
   [wait])
+
+(defendpoint get-current-application-information! nil
+  "Returns  a promise containing the bot's OAuth2 application info."
+  []
+  [])